+++ src/lisp_main.go
@@ -1,0 +1,13 @@
+/*
+   Go lang 3rd study program.
+   This is prototype program mini scheme subset.
+
+   hidekuno@gmail.com
+*/
+package main
+
+// Main
+func main() {
+	build_func()
+	do_interactive()
+}
